+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Ensure pool asynchronous resource properly track tasks execution
+
 ## [3.0.10] - 2023-12-08
 
 ### Changed